+++ components/components/componentsListView/ComponentsViewHomePage.tsx
@@ -1,11 +1,12 @@
 import { Block } from "baseui/block"
 import * as React from "react";
-import { Page } from "../../constants";
+import { Page } from "../../../constants";
 import { ComponentList } from "./ComponentList";
-import { componentsListView, container, content, nav, navbarWrapper } from "./componentStyles";
-import { CreateComponentButton } from "./createComponentButton";
-import { Header } from "./header"
+import { componentsListView, container, content, nav, navbarWrapper } from "../overrides/componentListStyles";
+import { CreateComponentButton } from "./CreateComponentButton";
+
 import { NavBarMenuItem } from "./NavBarMenuItem";
+import { Header } from "./Header";
 
 export const ComponentsViewHomePage: React.FC = () => {
   const [activePage, setPage] = React.useState<Page>(Page.Active); // variable indicating the current selected navbar menu-item
