+++ components/components/componentsListView/Header.tsx
@@ -1,12 +1,8 @@
 import { Block } from "baseui/block";
 import * as React from "react";
-import {
-  headerBar,
-  headerBarLeftWrapper,
-  headerBarBackIcon,
-} from "../incidents/list/overrides/navStyles";
 
 import { FiArrowLeft } from "react-icons/fi";
+import { headerBar, headerBarBackIcon, headerBarLeftWrapper } from "../../incidents/list/overrides/navStyles";
 
 interface Props {
   headerText: string; // text to render in the header
