+++ components/components/internal/Form/ComponentName.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { FormControl } from "baseui/form-control";
 import { Input } from "baseui/input";
 
-import { inputNameStyle } from "./componentCreationStyles";
+import { inputNameStyle } from "../../overrides/componentFormStyles";
 
 export const ComponentName = React.memo((props: any) => {
   return(
